Create one handleInputChange for controlled components in CreateRecipeDialogBox.
I had a different event-handler for each onChange event of the
controlled components in the CreateRecipeDialogBox, but it makes
more sense to add a 'name' attribute to each 'input' element that
corresponds to the state that the input represents, then just use
the es6 computed property syntax to generalize the event handlers
for all of the inputs. Neat trick!