Refactor the way in which recipes are deleted.
When the user clicks the 'delete recipe' button, I originally
passed a reference to the recipe to be deleted up the
component tree to the App component and filtered the 'recipes'
state by the name of the recipe that needed to be delted. I changed
this to use the 'id' of the recipe, since this is ensured to be
unique.