Change conditional rendering logic of CreateRecipeDialogBox and GreyOutBox.
I'm still not sure which pattern I prefer, but originally I was
passing a prop down to CreateRecipeDialogBox and GreyOutBox that
would determine whether the component rendered or returned null. This
looked clearer to me, but I've been working through tutorials and
they seem to prefer conditionally setting a variable that will hold
null or the component, so I'm refactoring to try that method out.